requirements.txt
Flask==2.3.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Jinja2==3.1.2
SQLAlchemy==2.0.21
Werkzeug==2.3.0
WTForms==3.0.1
Flask-session
flask_jwt_extended